+++ builtin/audit/file/backend_test.go
@@ -4,7 +4,6 @@
 package file
 
 import (
-	"context"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -13,7 +12,6 @@ import (
 	"github.com/hashicorp/eventlogger"
 	"github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/vault/audit"
-	"github.com/hashicorp/vault/helper/testhelpers/corehelpers"
 	"github.com/hashicorp/vault/internal/observability/event"
 	"github.com/hashicorp/vault/sdk/helper/salt"
 	"github.com/hashicorp/vault/sdk/logical"
@@ -41,7 +39,7 @@ func TestAuditFile_fileModeNew(t *testing.T) {
 		SaltView:   &logical.InmemStorage{},
 		Logger:     hclog.NewNullLogger(),
 	}
-	_, err = Factory(context.Background(), backendConfig, &corehelpers.NoopHeaderFormatter{})
+	_, err = Factory(backendConfig, &audit.NoopHeaderFormatter{})
 	require.NoError(t, err)
 
 	info, err := os.Stat(file)
@@ -74,7 +72,7 @@ func TestAuditFile_fileModeExisting(t *testing.T) {
 		Logger:     hclog.NewNullLogger(),
 	}
 
-	_, err = Factory(context.Background(), backendConfig, &corehelpers.NoopHeaderFormatter{})
+	_, err = Factory(backendConfig, &audit.NoopHeaderFormatter{})
 	require.NoError(t, err)
 
 	info, err := os.Stat(f.Name())
@@ -108,7 +106,7 @@ func TestAuditFile_fileMode0000(t *testing.T) {
 		Logger:     hclog.NewNullLogger(),
 	}
 
-	_, err = Factory(context.Background(), backendConfig, &corehelpers.NoopHeaderFormatter{})
+	_, err = Factory(backendConfig, &audit.NoopHeaderFormatter{})
 	require.NoError(t, err)
 
 	info, err := os.Stat(f.Name())
@@ -137,7 +135,7 @@ func TestAuditFile_EventLogger_fileModeNew(t *testing.T) {
 		Logger:     hclog.NewNullLogger(),
 	}
 
-	_, err = Factory(context.Background(), backendConfig, &corehelpers.NoopHeaderFormatter{})
+	_, err = Factory(backendConfig, &audit.NoopHeaderFormatter{})
 	require.NoError(t, err)
 
 	info, err := os.Stat(file)
@@ -243,7 +241,7 @@ func TestBackend_newFormatterConfig(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			got, err := newFormatterConfig(&corehelpers.NoopHeaderFormatter{}, tc.config)
+			got, err := newFormatterConfig(&audit.NoopHeaderFormatter{}, tc.config)
 			if tc.wantErr {
 				require.Error(t, err)
 				require.EqualError(t, err, tc.expectedMessage)
@@ -270,7 +268,7 @@ func TestBackend_configureFormatterNode(t *testing.T) {
 		nodeMap:    map[eventlogger.NodeID]eventlogger.Node{},
 	}
 
-	formatConfig, err := audit.NewFormatterConfig(&corehelpers.NoopHeaderFormatter{})
+	formatConfig, err := audit.NewFormatterConfig(&audit.NoopHeaderFormatter{})
 	require.NoError(t, err)
 
 	err = b.configureFormatterNode("juan", formatConfig, hclog.NewNullLogger())
@@ -425,8 +423,6 @@ func TestBackend_configureSinkNode(t *testing.T) {
 func TestBackend_Factory_Conf(t *testing.T) {
 	t.Parallel()
 
-	ctx := context.Background()
-
 	tests := map[string]struct {
 		backendConfig        *audit.BackendConfig
 		isErrorExpected      bool
@@ -493,7 +489,7 @@ func TestBackend_Factory_Conf(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			be, err := Factory(ctx, tc.backendConfig, &corehelpers.NoopHeaderFormatter{})
+			be, err := Factory(tc.backendConfig, &audit.NoopHeaderFormatter{})
 
 			switch {
 			case tc.isErrorExpected:
@@ -511,8 +507,6 @@ func TestBackend_Factory_Conf(t *testing.T) {
 // and set correctly, then exposed via the interface method IsFallback().
 func TestBackend_IsFallback(t *testing.T) {
 	t.Parallel()
-
-	ctx := context.Background()
 
 	tests := map[string]struct {
 		backendConfig      *audit.BackendConfig
@@ -552,7 +546,7 @@ func TestBackend_IsFallback(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			be, err := Factory(ctx, tc.backendConfig, &corehelpers.NoopHeaderFormatter{})
+			be, err := Factory(tc.backendConfig, &audit.NoopHeaderFormatter{})
 			require.NoError(t, err)
 			require.NotNil(t, be)
 			require.Equal(t, tc.isFallbackExpected, be.IsFallback())
